# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectedAccountResponse": {
        "properties": {
          "authorization_lapsed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorization Lapsed At"
          },
          "authorized_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized At"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "rule_count": {
            "title": "Rule Count",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "email",
          "provider",
          "avatar_url",
          "state",
          "rule_count",
          "authorized_at",
          "authorization_lapsed_at"
        ],
        "title": "ConnectedAccountResponse",
        "type": "object"
      },
...
      "DiscoveredCalendarResponse": {
        "properties": {
          "access_role": {
            "title": "Access Role",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "primary": {
            "title": "Primary",
            "type": "boolean"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "writable": {
            "title": "Writable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "summary",
          "access_role",
          "writable",
          "primary"
        ],
        "title": "DiscoveredCalendarResponse",
        "type": "object"
      },
...
      "GoogleAccountAccessResponse": {
        "properties": {
          "calendar_api": {
            "title": "Calendar Api",
            "type": "boolean"
          },
          "calendar_list_access": {
            "title": "Calendar List Access",
            "type": "boolean"
          },
          "calendars_visible": {
            "title": "Calendars Visible",
            "type": "integer"
          },
          "event_access": {
            "title": "Event Access",
            "type": "boolean"
          },
          "rules_resumed": {
            "title": "Rules Resumed",
            "type": "integer"
          },
          "writable_calendars": {
            "title": "Writable Calendars",
            "type": "integer"
          }
        },
        "required": [
          "calendar_api",
          "calendar_list_access",
          "event_access",
          "calendars_visible",
          "writable_calendars",
          "rules_resumed"
        ],
        "title": "GoogleAccountAccessResponse",
        "type": "object"
      },
      "GoogleConfigurationResponse": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "redirect_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect Uri"
          }
        },
        "required": [
          "configured",
          "redirect_uri"
        ],
        "title": "GoogleConfigurationResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Calendar Ghost",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "list_accounts_api_v1_accounts_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectedAccountResponse"
                  },
                  "title": "Response List Accounts Api V1 Accounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Accounts"
      }
    },
    "/api/v1/accounts/{account_id}": {
      "delete": {
        "operationId": "delete_account_api_v1_accounts__account_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Account"
      }
    },
    "/api/v1/accounts/{account_id}/calendars": {
      "get": {
        "operationId": "discover_calendars_api_v1_accounts__account_id__calendars_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscoveredCalendarResponse"
                  },
                  "title": "Response Discover Calendars Api V1 Accounts  Account Id  Calendars Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Calendars"
      }
    },
    "/api/v1/accounts/{account_id}/disconnect": {
      "post": {
        "operationId": "disconnect_account_api_v1_accounts__account_id__disconnect_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect Account"
      }
    },
    "/api/v1/accounts/{account_id}/verify": {
      "post": {
        "operationId": "verify_account_access_api_v1_accounts__account_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAccountAccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Account Access"
      }
    },
...
    "/api/v1/google/configuration": {
      "get": {
        "operationId": "google_configuration_api_v1_google_configuration_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleConfigurationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Configuration"
      }
    },
...
    "/api/v1/oauth/google/start": {
      "get": {
        "description": "Start Google's consent for the signed-in User; reauthorizing a known `account` suggests\nits email to Google.",
        "operationId": "start_google_oauth_api_v1_oauth_google_start_get",
        "parameters": [
          {
            "in": "query",
            "name": "account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Google Oauth"
      }
    },
...
  }
}